requirements.txt
asgiref==3.3.1
certifi==2020.12.5
chardet==4.0.0
Django==3.1.7
djangorestframework==3.12.4
djangorestframework-appapi==0.0.3
djangorestframework-jwt==1.11.0
idna==2.10
Naked==0.1.31
pycryptodome==3.10.1
PyJWT==1.7.1
PyMySQL==1.0.2
pytz==2021.1
PyYAML==5.4.1
requests==2.25.1
shellescape==3.8.1
sqlparse==0.4.1
urllib3==1.26.4
